Improve web ui in case of resizing
