+++ lib/views/home/home.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:yourshift_site/views/home/langind_page.dart';
+import 'package:yourshift_site/views/home/landpage/langind_page.dart';
 
 class HomeView extends StatelessWidget {
   const HomeView({Key? key, required this.title}) : super(key: key);
@@ -11,7 +11,11 @@ class HomeView extends StatelessWidget {
     return Scaffold(
       backgroundColor: Theme.of(context).backgroundColor,
       appBar: AppBar(
-        title: Text(title),
+        title: Text(title,
+            style: Theme.of(context)
+                .textTheme
+                .headline6!
+                .apply(color: Theme.of(context).textTheme.bodyText1!.color)),
         centerTitle: false,
         primary: true,
         elevation: 2,
