+++ pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
     sdk: flutter
   bottom_navy_bar: ^6.0.0
   url_launcher: ^6.0.12
+  font_awesome_flutter: ^9.2.0
 
 dev_dependencies:
   flutter_test:
