+++ web/index.html
@@ -39,6 +39,7 @@
   <title>Shift You</title>
   <link rel="manifest" href="manifest.json">
   <link rel="stylesheet" type="text/css" href="styles.css">
+
 </head>
 <body>
   <div class="center">
@@ -52,6 +53,14 @@
   <!-- This script installs service_worker.js to provide PWA functionality to
        application. For more information, see:
        https://developers.google.com/web/fundamentals/primers/service-workers -->
+  <!-- Global site tag (gtag.js) - Google Analytics -->
+  <script async src="https://www.googletagmanager.com/gtag/js?id=G-ABCDEFGHIJK"></script>
+  <script>
+    window.dataLayer = window.dataLayer || [];
+    function gtag(){dataLayer.push( arguments);}
+    gtag('js', new Date());
+    gtag('config', 'G-E0X8YETE09');
+  </script>
   <script>
     var serviceWorkerVersion = null;
     var scriptLoaded = false;
